% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_O_E_CUSUM.R
\name{mortality_O_E_CUSUM}
\alias{mortality_O_E_CUSUM}
\title{plots the mortality O-E CUSUM for each facility/center strata}
\usage{
mortality_O_E_CUSUM(
  data,
  startdate,
  enddate,
  HPT1,
  HPT2,
  year_event_rate,
  h_val_path,
  rho
)
}
\arguments{
\item{data}{The data user put in}

\item{startdate}{The calendar time of start date of the follow up (0 point of the time scale)}

\item{enddate}{The calendar time of start date of the follow up}

\item{HPT1}{The H_1 of deterioration}

\item{HPT2}{The H_1 of improvement}

\item{year_event_rate}{The risk}

\item{h_val_path}{Path of control limits for different sample sizes}

\item{rho}{Parameter to control the restart}
}
\value{
the signal points and not signal points
}
\description{
The function takes the information for survival analysis and returns the signal situation and plots
}
\examples{
mortality_O_E_CUSUM(data,startdate,enddate,HPT1,HPT2,year_event_rate,h_val_path,rho)
#Output: graphs
}
